fetch data by geo hash
