+++ routers/request/patient_detail_params.go
@@ -1,0 +1,1 @@
+package request
